% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{FilesToDESeqObj}
\alias{FilesToDESeqObj}
\title{FilesToDESeqObj}
\usage{
FilesToDESeqObj(sampleFile, countFile, dataType, session)
}
\arguments{
\item{sampleFile}{character - name of the sample file}

\item{countFile}{character - name of the count file}
}
\value{
a DESeqDataSet object
}
\description{
\code{FilesToDESeqObj} returns a DESeqDataSet object produced from the supplied sample and count files
}
\details{
The first column of the sample file must match the sample names in the column names of the count file
   The sample file must also have a column labelled "condition" for the DESeq2 design formula.
   If the samples file contains a column labelled "sampleName" the samples will be renamed.
}
\examples{
FilesToDESeqObj( 'samples.txt', 'all.tsv' )

}
